# My previous solution as a GraalVM native binary