--- a/calculate_average_kaufco.sh
+++ b/calculate_average_kaufco.sh
@@ -15,7 +15,7 @@
 #  limitations under the License.
 #
 
-JAVA_OPTS=""
+JAVA_OPTS="--enable-preview"
 java $JAVA_OPTS --class-path target/average-1.0.0-SNAPSHOT.jar dev.morling.onebrc.CalculateAverage_kaufco > java_output.log 2>&1 &
 
 # Capture the PID of the Java process