--- a/calculate_average_kaufco.sh
+++ b/calculate_average_kaufco.sh
@@ -15,5 +15,11 @@
 #  limitations under the License.
 #
 
-JAVA_OPTS="--enable-preview"
-java $JAVA_OPTS --class-path target/average-1.0.0-SNAPSHOT.jar dev.morling.onebrc.CalculateAverage_kaufco
+if [ -f target/CalculateAverage_kaufco_image ]; then
+    echo "Picking up existing native image 'target/CalculateAverage_kaufco_image', delete the file to select JVM mode." 1>&2
+    target/CalculateAverage_kaufco_image
+else
+    JAVA_OPTS="--enable-preview"
+    echo "Chosing to run the app in JVM mode as no native image was found, use prepare_kaufco.sh to generate." 1>&2
+    java $JAVA_OPTS --class-path target/average-1.0.0-SNAPSHOT.jar dev.morling.onebrc.CalculateAverage_kaufco
+fi