#### Changes as compared to PR #527:

- now using a static binary generated using runtime profiling data
- **new runtime: 0.96s** on my machine

#### Check List:
- [x] Tests pass (`./test.sh <username>` shows no differences between expected and actual outputs)
- [x] All formatting changes by the build are committed
- [x] Your launch script is named `calculate_average_<username>.sh` (make sure to match casing of your GH user name) and is executable
- [x] Output matches that of `calculate_average_baseline.sh`
- [x] For new entries, or after substantial changes: When implementing custom hash structures, please point to where you deal with hash collisions (line number)

#### Benchmark:

* Execution time: 0.96s
* Execution time of reference implementation:
  * baseline: 2m58.42s
  * merykitty: 2.11s
  * thomaswue: 1.35s

#### My test environment:

* AMD Ryzen 9 3950x (16 Cores)
* 32GB RAM
* an SSD (no RAMdisk used, but after first run the file is in the file system cache)
* Ubuntu 23.04
* GraalVM 22.0.1

<!--
Thanks for your submission. Please go through the checklist above before submitting your pull request.
Use [x] to mark that the item has been completed.

Due to the large number of entries created so far,
please submit only entries that are you are expecting to run in 10 seconds or less on the evaluation machine.

Please make sure that you have followed the defined rules (https://github.com/gunnarmorling/1brc?tab=readme-ov-file#rules-and-limits).
-->