@kaufco 2024-01-23
No, I would just rebase. This PR just changes the `calculate_average_kaufco.sh` and adds the `prepare_kaufco.sh`, but no changes on the Java file at all. Maybe its better to have 2 PRs for it, but if you prefer, we can also have all in this one.